% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHO2019Risk.R
\name{WHO_CVD_Risk_2019}
\alias{WHO_CVD_Risk_2019}
\title{Calculating 2019 WHO 10-year CVD risk scores}
\usage{
WHO_CVD_Risk_2019(age, gdr, smk, sbp, dm, chl, subregion)
}
\arguments{
\item{age}{age (years)}

\item{gdr}{gender (1 = male, 0 = female)}

\item{smk}{smoking status (1 = current smoker, 0 = non-smoker)}

\item{sbp}{systolic blood pressure (mmHg)}

\item{dm}{diabetes status (1 = diabetic, 0 = non-diabetic)}

\item{chl}{total cholesterol (mmol/L)}

\item{subregion}{epidemiolgical subregion (character value, one of: "N_AFR_ME", "CSS_AFR", "ESS_AFR", "SSS_AFR", "WSS_AFR", "SN_LAT_AME", "HI_N_AME", "CAR", "AND_LAT_AME", "CEN_LAT_AME", "TRO_LAT_AME", "EAS_ASI", "SOU_ASI", "SE_ASI", "CEN_ASI", "HI_ASI_PAC", "WES_EUR", "CEN_EUR", "EAS_EUR", "OCE", "AUS")}
}
\value{
A vector of WHO risk scores.
}
\description{
This function calculates 2019 WHO 10-year CVD risk scores for any epidemiolgical subregion of the world.
}
\details{
This function calculates the updated 2019 WHO 10-year CVD risk scores for any epidemiological subregion based on the values of six paramters passed to it (age, gender, smoking status, systolic blood pressure, diabetes status, and total cholesterol).

Subregions:

  "N_AFR_ME": North Africa and Middle East

  "CSS_AFR": Central Sub-Saharan Africa

  "ESS_AFR": Eastern Sub-Saharan Africa

  "SSS_AFR": Southern Sub-Saharan Africa

  "WSS_AFR": Western Sub-Saharan Africa

  "SN_LAT_AME": Southern Latin America

  "HI_N_AME": High-income North America

  "CAR": Caribbean

  "AND_LAT_AME": Andean Latin America

  "CEN_LAT_AME": Central Latin America

  "TRO_LAT_AME": Tropical Latin America

  "EAS_ASI": East Asia

  "SOU_ASI": South Asia

  "SE_ASI": Southeast Asia

  "CEN_ASI": Central Asia

  "HI_ASI_PAC": High-income Asia Pacific

  "WES_EUR": Western Europe

  "CEN_EUR": Central Europe

  "EAS_EUR": Eastern Europe

  "OCE": Oceania

  "AUS": Australasia
}
\references{
The WHO CVD Risk Chart Working Group. (2019). World Health
Organization cardiovascular disease risk charts: revised models to
estimate risk in 21 global regions.
*Lancet Glob Health* **(published online Sept 2.)**.
}
\author{
Rodrigo Citton P. dos Reis, Dylan R.J. Collins, Joseph Lee, Niklas Bobrovtiz, Constantinos Koshiaris, Alison Ward, Carl Heneghan
}
