% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHOISHRisk.R
\name{WHO_ISH_Risk}
\alias{WHO_ISH_Risk}
\title{Calculating WHO/ISH Risk Scores}
\usage{
WHO_ISH_Risk(age, gdr, smk, sbp, dm, chl, subregion)
}
\arguments{
\item{age}{age (years)}

\item{gdr}{gender (1 = male, 0 = female)}

\item{smk}{smoking status (1 = current smoker, 0 = non-smoker)}

\item{sbp}{systolic blood pressure (mmHg)}

\item{dm}{diabetes status (1 = diabetic, 0 = non-diabetic)}

\item{chl}{total cholesterol (mmol/L)}

\item{subregion}{epidemiolgical subregion (character value, one of: “AFR_D”, “AFR_E”, “AMR_A”, “AMR_B”, “AMR_D”,“EMR_B”,“EMR_D”, “EUR_A”,“EUR_B”, EUR_C”, “SEAR_B”, “SEAR_D”, “WPR_A”, “WPR_B”)}
}
\value{
A vector of WHO/ISH risk scores.
}
\description{
This function calculates WHO/ISH risk scores for any epidemiolgical subregion of the world.
}
\details{
This function calculates WHO/ISH risk scores for any epidemiological subregion based on the values of six paramters passed to it (age, gender, smoking status, systolic blood pressure, diabetes status, and total cholesterol).
}
\author{
Dylan R.J. Collins, Joseph Lee, Niklas Bobrovtiz, Constantinos Koshiaris, Alison Ward, Carl Heneghan
}

